textbook/rag-backend/src/services/translation_service.py
from typing import Dict, Any, Optional
from src.models.translation import Translation
import googletrans  # In a real app, we'd use proper Google Translate API

class TranslationService:
    """
    Service for handling content translation between languages
    """
    
    def __init__(self):
        # In a real implementation, this would connect to a translation API
        # For now, using a simple translation mapping for demo purposes
        self.translation_cache = {}
        
        # Sample translations for key terms (in a real app, this would be much more comprehensive)
        self.translation_map = {
            "English": {
                "en": {},
                "ur": {
                    "Introduction to ROS 2": "ROS 2 کا تعارف",
                    "Vision-Language-Action Systems": "رویت-زبان-عمل کے نظام",
                    "Physical AI & Humanoid Robotics": "جسمانی مصنوعی ذہانت اور انسان نما روبوٹکس",
                    "Chapter": "باب",
                    "Module": "ماڈیول"
                }
            }
        }
    
    def translate_content(self, 
                         content_id: str, 
                         original_content: Dict[str, Any], 
                         target_language: str,
                         source_type: str) -> Optional[Translation]:
        """
        Translate content to the target language
        """
        # Check if translation exists in cache
        cache_key = f"{content_id}_{target_language}"
        if cache_key in self.translation_cache:
            return self.translation_cache[cache_key]
        
        # In a real implementation, this would call a translation API
        # For example: Google Cloud Translation API, Azure Translator API, etc.
        
        # For now, implement a basic translation based on our sample translation map
        translated_content = {}
        
        if isinstance(original_content, dict):
            for key, value in original_content.items():
                if isinstance(value, str):
                    # Try to find a translation in our map
                    translated_value = self.translate_text(value, target_language)
                    translated_content[key] = translated_value
                else:
                    translated_content[key] = value
        elif isinstance(original_content, str):
            translated_content = self.translate_text(original_content, target_language)
        else:
            translated_content = original_content  # For other types, just pass through
        
        # Create translation object
        translation = Translation(
            id=f"tr_{content_id}_{target_language}",
            originalContentId=content_id,
            originalContentType=source_type,
            language=target_language,
            translatedContent=translated_content
        )
        
        # Cache the translation
        self.translation_cache[cache_key] = translation
        
        return translation
    
    def translate_text(self, text: str, target_language: str) -> str:
        """
        Translate a text string to the target language
        """
        # In a real app, this would call an actual translation API
        if target_language.lower() == "ur":
            # Use our simple translation map for demonstration
            for english_term, urdu_term in self.translation_map["English"]["ur"].items():
                if english_term in text:
                    text = text.replace(english_term, urdu_term)
            
            # For demonstration purposes, return a message indicating the text has been translated
            return f"[TRANSLATED TO URDU]: {text}"
        else:
            return text  # Return original for other languages (demo purposes)
    
    def get_cached_translation(self, content_id: str, target_language: str) -> Optional[Translation]:
        """
        Retrieve a cached translation
        """
        cache_key = f"{content_id}_{target_language}"
        return self.translation_cache.get(cache_key)

# Create a global instance
translation_service = TranslationService()